requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.34.0
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
